third_party/WebKit/LayoutTests/http/tests/devtools/elements/styles-2/inactive-properties.html

<html>
<head>
<script src="../../../inspector/inspector-test.js"></script>
<script src="../../../inspector/elements-test.js"></script>
<script>

function test() {
  ElementsTestRunner.selectNodeAndWaitForStylesWithComputed('inspected', step1);

  function step1() {
    ElementsTestRunner.dumpSelectedElementStyles(false, false);
    TestRunner.completeTest();
  }
}

</script>
<style>
#inspected {
    text-align: left;
    text-align: bar;
    text-align: right;
}
</style>
</head>

<body onload="runTest()">
<p>
Tests that effectively inactive properties are displayed correctly in the sidebar.
</p>

<div id="container">
    <div id="inspected" align="left">Test</div>
</div>

</body>
</html>
