third_party/WebKit/LayoutTests/http/tests/devtools/elements/styles-2/inactive-properties-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that effectively inactive properties are displayed correctly in the sidebar.

display: block;
    block - div user agent stylesheet
text-align: right;
    right - #inspected <style>…</style>
    OVERLOADED -webkit-left - element.style
[expanded] 
element.style { ()

[expanded] 
#inspected { (<style>…</style>)
/-- overloaded --/     text-align: left;
/-- overloaded --/     text-align: bar;
    text-align: right;

[expanded] 
div[Attributes Style] { ()
/-- overloaded --/     text-align: -webkit-left;

[expanded] 
div { (user agent stylesheet)
    display: block;
